requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
bs4==0.0.1
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
six==1.10.0
xlrd==1.0.0
XlsxWriter==0.9.6
